#33 refactor

+++ src/ConnectKafkaFake.cpp
@@ -32,7 +32,7 @@ int ConnectKafkaFake::getNumberOfTopicPartitions(std::string TopicName) {
 void ConnectKafkaFake::subscribeAtOffset(int64_t Offset,
                                          std::string TopicName) {}
 
-void ConnectKafkaFake::subscribeToLastNMessages(int32_t NMessages,
+void ConnectKafkaFake::subscribeToLastNMessages(int64_t NMessages,
                                                 const std::string &TopicName,
                                                 int Partition) {}
